12272016 changed some db types

+++ src/main/java/com/dalonedrau/entities/avalon/Advantage.java
@@ -6,7 +6,7 @@ import com.dalonedrau.schemacreator.Annotations.VarChar;
 public class Advantage {
     private String description;
     /** advantage flag. */
-    private int flag;
+    private long flag;
     @Unique
     @VarChar(length = 40)
     private String name;
